+++ src/security.py
@@ -1,0 +1,15 @@
+import os
+
+from fastapi import Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordBearer
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
+
+def api_key_auth(
+    api_key: str = Depends(oauth2_scheme),
+    api_keys: str = os.environ.get("API_KEY", "a"),
+):
+    if api_key not in api_keys:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED, detail="Forbidden"
+        )
